third_party/blink/web_tests/platform/generic/printing/page-break-orphans-expected.txt

Test for orphans

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS: page number of "page1-1" is 1
PASS: page number of "page1-2" is 1
PASS: page number of "page3" is 3
PASS: page number of "page4" is 4
PASS: page number of "page5-1" is 5
PASS: page number of "page5-2" is 5
All tests passed

PASS successfullyParsed is true

TEST COMPLETE
